refactor(types): popover
